re-run a xhigh review and fix more nits
