+++ compiler/plc_lowering/src/lib.rs
@@ -6,6 +6,7 @@ use plc_ast::{ast::AstNode, provider::IdProvider};
 
 pub mod array_lowering;
 pub mod control_statement;
+pub(crate) mod helper;
 pub mod inheritance;
 pub mod initializer;
 pub mod retain;
